requirements.txt
flask
bcrypt
argon2-cffi
pyotp
requests
pandas
matplotlib
seaborn